pyproject.toml
[tool.poetry]
name = "tk-tagger"
version = "0.1.0"
description = ""
authors = ["ggzor <30713864+ggzor@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "3.9"
Pillow = "^8.3.2"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
pynvim = "^0.4.3"
pylint = "^2.11.1"
rope = "^0.20.1"
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
